+++ codex-rs/core/BUILD.bazel
@@ -49,10 +49,3 @@ codex_rust_crate(
         "//codex-rs/windows-sandbox-rs:codex-windows-sandbox-setup",
     ],
 )
-
-alias(
-    name = "wine-remote-env-test-driver",
-    testonly = True,
-    actual = ":core-all-test",
-    visibility = ["//codex-rs/exec-server/tests/foreign-os:__pkg__"],
-)
